+++ qcar_updated/launch/qcar_nav2.launch.py
@@ -3,7 +3,7 @@ from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
 from launch.actions import IncludeLaunchDescription, DeclareLaunchArgument
 from launch.actions import TimerAction
-from launch.conditions import IfCondition
+from launch.conditions import IfCondition, UnlessCondition
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration
 from nav2_common.launch import RewrittenYaml
@@ -36,7 +36,7 @@ def generate_launch_description():
     return LaunchDescription([
         DeclareLaunchArgument(
             'launch_sim',
-            default_value='true',
+            default_value='false',
             description='Bring up Gazebo + robot_state_publisher + controllers as well'
         ),
 
@@ -45,6 +45,14 @@ def generate_launch_description():
                 os.path.join(pkg, 'launch', 'qcar_updated.launch.py')
             ),
             condition=IfCondition(launch_sim)
+        ),
+
+        # real-hardware equivalent when launch_sim:=false
+        IncludeLaunchDescription(
+            PythonLaunchDescriptionSource(
+                os.path.join(pkg, 'launch', 'qcar_visualize.launch.py')
+            ),
+            condition=UnlessCondition(launch_sim)
         ),
 
         # Delay everything below by 3 seconds
@@ -57,7 +65,7 @@ def generate_launch_description():
                         os.path.join(nav2_bringup_dir, 'launch', 'bringup_launch.py')
                     ),
                     launch_arguments={
-                        'use_sim_time': 'true',
+                        'use_sim_time': 'False',
                         'params_file': configured_nav2_params,
                         'map': map_file
                     }.items()